requirements.txt
amqp==5.1.1
anyio==3.6.2
bcrypt==4.0.1
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.2
dnspython==2.3.0
ecdsa==0.18.0
fastapi==0.95.0
h11==0.14.0
httptools==0.5.0
idna==3.4
kombu==5.2.4
motor==3.1.1
networkx==3.0
numpy==1.24.2
passlib==1.7.4
prompt-toolkit==3.0.38
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.6
pymongo==4.3.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2022.7.1
PyYAML==6.0
rsa==4.9
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
uvloop==0.17.0
vine==5.0.0
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
